Add functions for users to retrieve weather infos.
Revise the event listener from button click to form submit for a more
versatile manipulation of form submission.
Created the clearForm function to clear the city input every submission.
Created the postCard function to receive weather data, create cards and
post it on the website.
Created the getWeather function to get the weather data of the desired
city.
Created the facilitateSubmission function to manage the user submission,
getting the weather and posting the card.